fix(docs): remove pro banner on examples routes
